Updated code
Removed some print statements for cleaner CLI output. Added a condition check to see if the sum of the mods is greater than 0. If the mod sum is 0 and there is a free value, do nothing. There are instances where there could be a general free amount but no default mod, mostly in Cheese Choice options.